requirements.txt
langchain-community>=0.0.20
langchain-openai>=0.0.20
python-dotenv>=1.0.0
beautifulsoup4>=4.13.0
requests>=2.31.0
pdfkit>=1.0.0
pytest>=7.0.0